refactor: split db commands into separate files
Additionally we have changed the signature of the InstallDatabase step
to take the logger as the last parameter.

sc-18443
